apps/api/requirements.txt
Django==6.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
psycopg[binary]==3.3.4
python-dotenv==1.0.1
Pillow==11.0.0
django-cors-headers==4.3.1
